Porting changes from branch pk/155791
Changes to supposedly fix the bug
155791 : Scaffolding needs to add the connection strings to
appSettings.json file when creating a new DbContext

However there seems to be a design issue or limitation for the above to
work.

Scaffolding loads the app with existing appsettings.json file and even if
scaffolding
changes the file contents on disk, the app is not reloaded at that point
and so running the app
to create a context instance is not working (it's not taking the new
contents of appsettings.json into consideration).

I am not sure how this can be fixed. Fowler is the best person to get some
ideas from.

The above code changes are all what is required for modification to code /
appsettings.files, I have tests covering that as well.